refactor(privacy): the SAR gather list is its own file
sar.go crossed the 500-line ceiling. The split is by the question each
half answers: sarsections.go is WHICH tables hold data about a subject and
the query that reads each (the compliance-critical half — adding or
dropping a source changes what the export owes), sar.go is how the export
runs (permission gate, transaction, row scan).

Three gates name the file their subject lives in. The PII coverage gate
says so in its own doc and is given the new file; the profile-field and
edge-reader waivers are repointed. All three fail loudly on a stale key
rather than passing quietly, which is why the move was safe to make.
