+++ backend/gates/edgereaders_test.go
@@ -131,7 +131,7 @@ var predicateEdgeReads = gatekit.Waive(map[string]string{
 // than the disclosure this change fixes: the sweep's correctness depends on
 // seeing every edge, and it runs as the system principal precisely so it does.
 var lifecycleEdgeReads = gatekit.Waive(map[string]string{
-	"internal/modules/privacy/sar.go":                                      "the subject-access export must enumerate every edge naming the data subject, or the export it produces is incomplete — a lawful-processing defect. Runs as the system principal on a request a human already authorised. The cost is that this path reads every edge of one subject with no per-caller gate",
+	"internal/modules/privacy/sarsections.go":                              "the subject-access export must enumerate every edge naming the data subject, or the export it produces is incomplete — a lawful-processing defect. Runs as the system principal on a request a human already authorised. The cost is that this path reads every edge of one subject with no per-caller gate",
 	"internal/modules/privacy/erasure_graph.go":                            "the erasure walk must reach every edge naming the subject or it leaves data behind that was ordered deleted. Respecting a caller's grants here would under-delete. The cost is unlimited edge reach for the sweep, bounded by its system principal",
 	"internal/modules/privacy/retentionselectors.go":                       "the selector excludes rows still anchored by a live edge, so an edge it cannot see would be deleted while still referenced. The cost is unlimited edge reach in a predicate whose output is a delete bound",
 	"internal/modules/people/relationship.go:lockPersonForEmployment":      "the row lock a stakeholder or employment write takes before it changes the edge, called only from CreateRelationship and UpdateRelationship, both of which ask the relationship create and update grants at their own entry. The cost is that the lock trusts its callers rather than re-asking",
