+++ backend/gates/profilefieldreaders_test.go
@@ -84,7 +84,7 @@ var selectsValues = regexp.MustCompile(
 var profileFieldValueReaders = gatekit.Waive(map[string]string{
 	"internal/modules/people/mergerelink.go":      "the merge copies each surviving evidence row from the merged-away person to the survivor, value and provenance together, and deletes the source rows. It serves nobody: the values move between two records and are read back through person360 afterwards like any other, verdict and all. Overlaying here would write the OVERLAID value into the survivor's row and destroy the distinction between what the machine asserted and what a human decided about it — the copy must be faithful precisely because the ledger is consulted later. The cost is that a verdict keyed to the merged-away person's id does not follow its value across, which is a gap in the merge rather than in this reader",
 	"internal/modules/people/orgnamepromotion.go": "RATIFIED PENDING A DECISION, see issue #2319. The corroborated org-name promotion reads the org_name signature values to count how many people at an organization state the same employer name. It does not show them as fact — it ranks a claim and, when signatures alone are the only agreement, stages a 🟡 proposal for a human. But it does not consult the ledger either, so an org_name a human already REJECTED still corroborates. Whether a verdict about one person's signature should veto a rename of their employer is a product question with an argument on each side, which is why it is filed rather than decided here. The cost, stated plainly: until that is answered, a rejected claim carries weight it may not deserve",
-	"internal/modules/privacy/sar.go":             "the Article 15 export owes the subject what this installation HOLDS, and it holds two facts: the value the machine asserted and the verdict the human recorded against it. It exports the stored columns here and ai_feedback as its own section beside them, so the subject sees both. Overlaying the verdict instead would hand them one merged value and conceal that an override exists — the opposite of what an export is for. It also cannot share person360's reader: privacy is a module and may not import compose (ADR-0054 §3). The cost is that the two statements must be corrected together when the column set changes, which is why they name each other at both sites",
+	"internal/modules/privacy/sarsections.go":     "the Article 15 export owes the subject what this installation HOLDS, and it holds two facts: the value the machine asserted and the verdict the human recorded against it. It exports the stored columns here and ai_feedback as its own section beside them, so the subject sees both. Overlaying the verdict instead would hand them one merged value and conceal that an override exists — the opposite of what an export is for. It also cannot share person360's reader: privacy is a module and may not import compose (ADR-0054 §3). The cost is that the two statements must be corrected together when the column set changes, which is why they name each other at both sites",
 })
 
 var profileFieldReaderScope = gatekit.Scope{
